decrating: fold in-tree registry logic into shipper-registry crate
Move the canonical 4791-LOC implementation from crates/shipper/src/registry.rs
INTO crates/shipper-registry/src/, split into:

- lib.rs (facade, top-level fns, constants)
- context.rs (canonical Registry-aware RegistryClient + tests + snapshots)
- http.rs (lightweight HttpRegistryClient with bare-url API, kept for
  shipper-engine-parallel; previously the entire public surface)

shipper-registry becomes the complete public registry client. shipper
depends on it via `pub use shipper_registry as registry;` re-export at
lib.rs, preserving the `shipper::registry::*` surface.

Delete the in-tree duplicate (registry.rs, 4791 LOC) and the shim
(registry_micro.rs, 239 LOC). Drop the `micro-registry` feature flag.
Make shipper-registry a non-optional dep of shipper.

Move insta snapshots:
- crates/shipper/src/snapshots/shipper__registry__tests__*.snap
  -> crates/shipper-registry/src/snapshots/shipper_registry__context__tests__*.snap
- Rename pre-existing snapshots from `shipper_registry__tests__*` to
  `shipper_registry__http__tests__*` to match new module path.

shipper-engine-parallel continues to use the bare-url API via
`use shipper_registry::HttpRegistryClient as RegistryClient;` (single-line
import change).

Update .github/workflows/ci.yml, templates/circleci-config.yml and
docs/architecture.md to remove `micro-registry` references.

Per docs/decrating-plan.md Phase 4 (registry special case).
