+++ crates/shipper-cli/Cargo.toml
@@ -38,7 +38,6 @@ micro-environment = ["shipper/micro-environment"]
 micro-storage = ["shipper/micro-storage"]
 micro-cargo = ["shipper/micro-cargo"]
 micro-plan = ["shipper/micro-plan"]
-micro-registry = ["shipper/micro-registry"]
 micro-webhook = ["shipper/micro-webhook"]
 micro-types = ["shipper/micro-types"]
 micro-config = ["shipper/micro-config"]
