+++ crates/shipper-registry/Cargo.toml
@@ -15,7 +15,10 @@ anyhow = "1.0"
 serde.workspace = true
 serde_json = "1.0"
 reqwest = { version = "0.13", features = ["blocking", "json", "rustls"] }
+chrono = { version = "0.4", features = ["serde"] }
+rand = { version = "0.10", features = ["std"] }
 shipper-sparse-index = { path = "../shipper-sparse-index" }
+shipper-types = { path = "../shipper-types" }
 
 [dev-dependencies]
 insta = { workspace = true }
