+++ crates/shipper/Cargo.toml
@@ -25,7 +25,7 @@ shipper-storage = { path = "../shipper-storage", optional = true }
 shipper-cargo = { path = "../shipper-cargo", optional = true }
 shipper-engine-parallel = { path = "../shipper-engine-parallel", optional = true }
 shipper-plan = { path = "../shipper-plan", optional = true }
-shipper-registry = { path = "../shipper-registry", optional = true }
+shipper-registry = { path = "../shipper-registry" }
 shipper-webhook = { path = "../shipper-webhook" }
 shipper-types = { path = "../shipper-types" }
 shipper-config = { path = "../shipper-config" }
@@ -68,12 +68,11 @@ micro-environment = ["dep:shipper-environment", "micro-types"]
 micro-storage = ["dep:shipper-storage"]
 micro-cargo = ["dep:shipper-cargo"]
 micro-plan = ["dep:shipper-plan", "micro-types"]
-micro-registry = ["dep:shipper-registry"]
 micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
 micro-config = ["micro-types", "micro-storage"]
-micro-parallel = ["micro-types", "micro-plan", "micro-registry", "dep:shipper-engine-parallel"]
-micro-all = ["micro-auth", "micro-git", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-registry", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
+micro-parallel = ["micro-types", "micro-plan", "dep:shipper-engine-parallel"]
+micro-all = ["micro-auth", "micro-git", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
 
 [lints]
 workspace = true
